+++ src/http/middleware/index.js
@@ -1,7 +1,6 @@
 let binary = require('./_binary-handler')
 let parse = require('./_parse')
 let _static = require('./_static-path')
-let _asd = require('./_services')
 let _fallback = require('./_fallback')
 let cors = require('./_cors')
 
@@ -14,10 +13,6 @@ module.exports = function loadMiddleware (app, { cwd, inventory, update }) {
 
   // Direct static asset delivery via /_static
   app.use(_static)
-
-  // Resource discovery
-  let asd = _asd.bind({}, inventory)
-  app.use(asd)
 
   // Route fallthrough to @proxy + ASAP
   let fallback = _fallback.bind({}, { cwd, inventory, update })